scripts/gis-material-update/process/requirements.txt
pandas
gtfs_kit
pyyaml
matplotlib
sqlalchemy
psycopg2-binary
geoalchemy2
pyjq
ipykernel
parse